Bit of a refactor after feature creep, prepping for much better designed --light and --superlight modes to get around circular deps, replacing the pretty crappy 'initial' double render approach
