+++ py_rust/src/render/debug.rs
@@ -1,8 +1,8 @@
-use crate::config::Config;
+use crate::state::State;
 
 #[derive(Debug, serde::Serialize)]
 pub struct Debug {
-    pub config: Config,
+    pub state: State,
     pub written: Vec<String>,
     pub identical: Vec<String>,
     pub matched_templates: Vec<String>,
