+++ py_rust/src/render/mod.rs
@@ -8,16 +8,13 @@ use minijinja::context;
 mod args_validate;
 mod debug;
 mod lockfile;
+mod mini_env;
 mod template;
 mod walker;
 pub use lockfile::hash_contents;
 pub use walker::get_template_matcher_rewrite_mapping;
 
-use crate::{
-    args::RenderCommand,
-    config::{self, Config},
-    prelude::*,
-};
+use crate::{args::RenderCommand, prelude::*, render::mini_env::new_mini_env, state::State};
 
 pub fn render(args: &crate::args::Args, render_args: &RenderCommand) -> Result<bool, Zerr> {
     args_validate::args_validate(render_args)?;
@@ -26,35 +23,43 @@ pub fn render(args: &crate::args::Args, render_args: &RenderCommand) -> Result<b
         self::lockfile::Lockfile::load(render_args.root.clone(), render_args.force)
     });
 
-    let mut conf = config::load(
-        args,
-        Some(render_args),
-        None,
-        // If newly created, should use cli initials if specified:
-        lockfile.newly_created,
-    )?;
+    let mut state = State::new(args)?;
 
-    // Need to run twice and rebuild config with real cli vars if initials used in the first conf build:
-    let (written, identical) = if conf.cli_initials_used {
+    // If newly created, should use cli initials if any have them:
+    let use_cli_initials = lockfile.newly_created
+        && state
+            .conf
+            .context
+            .cli
+            .iter()
+            .any(|(_, v)| v.initial.is_some());
+    state.load_all_vars(Some(render_args), use_cli_initials)?;
+
+    // Need to run twice and rebuild config with real cli vars if initials used in the first state build:
+    let (written, identical) = if use_cli_initials {
         warn!("Lockfile newly created/force updated and some cli vars have initials, will double render and use initials first time round.");
         // Conf from second as that has the real cli vars, template info from the first as the second will be inaccurate due to the first having run.
-        let (first_written, first_identical) = render_inner(&conf, render_args, &mut lockfile)?;
-        conf = config::load(args, Some(render_args), None, false)?;
-        render_inner(&conf, render_args, &mut lockfile)?;
+        let (first_written, first_identical) = render_inner(&state, render_args, &mut lockfile)?;
+
+        // Reload with real cli vars:
+        state.load_all_vars(Some(render_args), false)?;
+
+        // Re-render:
+        render_inner(&state, render_args, &mut lockfile)?;
         (first_written, first_identical)
     } else {
-        render_inner(&conf, render_args, &mut lockfile)?
+        render_inner(&state, render_args, &mut lockfile)?
     };
 
     // Run post-tasks:
-    conf.tasks.run_post(&conf)?;
+    state.conf.tasks.run_post(&state)?;
 
     timeit!("Syncing lockfile", { lockfile.sync() })?;
 
     // Write only when hidden cli flag --debug is set, to allow testing internals from python without having to setup custom interfaces:
     if render_args.debug {
         let debug = debug::Debug {
-            config: conf,
+            state: state.clone(),
             written: written
                 .iter()
                 .map(|t| t.out_path.display().to_string())
@@ -80,12 +85,18 @@ pub fn render(args: &crate::args::Args, render_args: &RenderCommand) -> Result<b
             .change_context(Zerr::InternalError)?;
     }
 
+    let num_tasks = state.conf.tasks.pre.len() + state.conf.tasks.post.len();
     println!(
-        "{} {} template{} written, {} identical. Lockfile {}. {} elapsed.",
+        "{} {} template{} written, {} identical.{} Lockfile {}. {} elapsed.",
         "zetch:".bold(),
         written.len(),
         if written.len() == 1 { "" } else { "s" },
         identical.len(),
+        if num_tasks > 0 {
+            format!("{} tasks run.", num_tasks).to_string()
+        } else {
+            "".to_string()
+        },
         if lockfile.modified {
             "modified"
         } else {
@@ -102,7 +113,7 @@ pub fn render(args: &crate::args::Args, render_args: &RenderCommand) -> Result<b
 }
 
 fn render_inner(
-    conf: &Config,
+    state: &State,
     render_args: &RenderCommand,
     lockfile: &mut self::lockfile::Lockfile,
 ) -> Result<
@@ -113,11 +124,11 @@ fn render_inner(
     Zerr,
 > {
     let walker = timeit!("Filesystem walker creation", {
-        self::walker::create(&render_args.root, conf)
+        self::walker::create(&render_args.root, state)
     })?;
 
     let templates = timeit!("Traversing filesystem & identifying templates", {
-        self::walker::find_templates(&render_args.root, walker, conf.matchers.as_slice())
+        self::walker::find_templates(&render_args.root, walker, state.conf.matchers.as_slice())
     })?;
 
     let mut identical = Vec::new();
@@ -126,7 +137,7 @@ fn render_inner(
     // Create the minijinja environment with the context.
     // A loader is set that can automatically load templates, this means it can load the main templates, and any other "includes" in user templates too.
     let env = timeit!("Creating rendering environment", {
-        conf.engine.create_minijinja_env(&render_args.root, conf)
+        new_mini_env(&render_args.root, state)
     })?;
 
     timeit!("Rendering templates & syncing files", {
